+++ .editorconfig
@@ -1,0 +1,16 @@
+root = true
+
+[*]
+indent_style = space
+
+indent_size = 4
+
+charset = utf-8
+
+trim_trailing_whitespace = true
+
+insert_final_newline = true
+
+end_of_line = lf
+
+max_line_length = off
