--- .env
@@ -1,2 +1,0 @@
-PORT=9000
-NODE_ENV=dev
