--- extra Ts/index.ts
@@ -1,28 +1,0 @@
-import { validateUsers } from "./utils/index";
-
-const users: IUsers[] =
-  [
-    {
-      traineeEmail: "satish1.patel@successive.tech",
-      reviewerEmail: "shubham1.jain@successive.tech"
-    },
-    {
-      traineeEmail: "satish2.patel@successive.tech",
-      reviewerEmail: "shubham2.jain@successive.tech"
-    },
-    {
-      traineeEmail: "satish13.patel@successive.tech",
-      reviewerEmail: "shubham13.jain@successive.tech"
-    },
-    {
-      traineeEmail: "satish14.patel@successive.tech",
-      reviewerEmail: "shubham14@.jain@successive.tech"
-    }
-  ];
-
-  validateUsers(users);
-  import {IUsers} from "./interfaces";
-
-  
-
-
