--- extra Ts/pattern/index.ts
@@ -1,4 +1,0 @@
-import diamond from './Diamond.js';
-import equilateral from './Equilateral.js';
-export {diamond};
-export {equilateral};
