--- extra Ts/utils/index.ts
@@ -1,5 +1,0 @@
-import {hasPermission} from './permissions.js';
-import {validateUsers} from './validation.js';
-export {hasPermission};
-export {validateUsers};
-//index.js
