+++ extraTs/pattern/index.ts
@@ -1,0 +1,3 @@
+import diamond from './Diamond';
+import equilateral from './Equilateral';
+export { diamond, equilateral };
