+++ extraTs/utils/index.ts
@@ -1,0 +1,4 @@
+import {hasPermission} from './permissions';
+import {validateUsers} from './validation';
+export {hasPermission};
+export {validateUsers};
