+++ extraTs/utils/validation.ts
@@ -1,12 +1,8 @@
-let checkemail = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+import { validateEmail } from './helpers';
+import { IUsers } from './extraTs/interface';
 let vldUser : String[]=[];
 let inVldUser : String[]=[];
-function validateEmail(email) 
-{
-  return checkemail.test(String(email).toLowerCase());
-}
-
-function validateUsers(user) 
+function validateUsers(user: IUsers)
 {
   user.forEach((element) =>
   {
@@ -26,5 +22,3 @@ console.log("Total Invalid users : ", inVldUser.length);
 console.log("Invalid users are : ", inVldUser);
 }
 export {validateUsers};
-export {validateEmail};
-
