+++ tsconfig.json
@@ -20,6 +20,6 @@
     },
     },
     "exclude": ["dist"],
-    "include": ["extra Ts"],
+    "include": ["extraTs"],
     "types": ["src/types"]
 }
